requirements.txt
langchain==0.1.0
langchain-groq==0.0.1
langgraph==0.0.20
pydantic==2.5.0
python-dotenv==1.0.0
streamlit==1.29.0
pillow==10.1.0
faiss-cpu==1.7.4
langchain-community==0.0.10
sentence-transformers==2.2.2